refactor: 사용자 입력 값에 대한 검증 로직 분리
