+++ src/App.js
@@ -1,10 +1,12 @@
 import { MissionUtils } from '@woowacourse/mission-utils';
 import View from './View.js';
+import Validate from './Validate.js';
 
 class App {
   constructor() {
     this.view = new View();
     this.view.infoPrint('숫자 야구 게임을 시작합니다.');
+    this.validate = new Validate();
   }
 
   async play() {
@@ -25,34 +27,6 @@ class App {
       }
     }
     return computer;
-  }
-
-  userPickValidation(value) {
-    if (value === undefined || value.length !== 3) {
-      throw new Error('[ERROR] 서로 다른 숫자 3개를 입력해주세요.');
-    }
-
-    const valueArr = [
-      ...new Set(
-        value
-          .split('')
-          .map((element) => +element)
-          .filter((element) => !isNaN(element))
-          .filter((element) => element > 0 && element < 10),
-      ),
-    ];
-    if (valueArr.length !== 3) {
-      throw new Error('[ERROR] 서로 다른 숫자 3개를 입력해주세요.');
-    }
-  }
-
-  userRestartValidation(value) {
-    if (!(value === '1' || value === '2'))
-      throw new Error('[ERROR] 새로 시작하려면 1, 종료하려면 2를 입력하세요.');
-  }
-
-  throwError(message) {
-    throw new Error(message);
   }
 
   judge(computer, user) {
@@ -80,7 +54,7 @@ class App {
     const computerNumber = this.computerPick();
     while (judgeResult !== '3스트라이크') {
       const userPickValue = await this.view.userInput('숫자를 입력해주세요 : ');
-      this.userPickValidation(userPickValue);
+      this.validate.userPickNumbers(userPickValue);
       const userNumber = userPickValue.split('').map((element) => +element);
       judgeResult = this.judge(computerNumber, userNumber);
       this.view.infoPrint(judgeResult);
@@ -89,7 +63,7 @@ class App {
     const userRestartValue = await this.view.userInput(
       `게임을 새로 시작하려면 1, 종료하려면 2를 입력하세요.\n`,
     );
-    this.userRestartValidation(userRestartValue);
+    this.validate.restartOrExit(userRestartValue);
     if (userRestartValue === '1') this.game();
     if (userRestartValue === '2') {
       this.view.infoPrint('숫자 야구 게임을 종료합니다.');
